# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/platform/plans": {
      "get": {
        "operationId": "SubscriptionPlansController_findAll",
        "parameters": [
          {
            "name": "all",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Plans"
        ]
      },
      "post": {
        "operationId": "SubscriptionPlansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Platform Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/plans/{id}": {
      "put": {
        "operationId": "SubscriptionPlansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SubscriptionPlansController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/organizations": {
      "post": {
        "operationId": "PlatformOrganizationsController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Platform Organizations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PlatformOrganizationsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Organizations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/organizations/{id}": {
      "get": {
        "operationId": "PlatformOrganizationsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Platform Organizations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "PlatformOrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Organizations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/organizations/{id}/suspend": {
      "put": {
        "operationId": "PlatformOrganizationsController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Organizations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/organizations/{id}/activate": {
      "put": {
        "operationId": "PlatformOrganizationsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Organizations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/organizations/{id}/subscription": {
      "put": {
        "operationId": "PlatformOrganizationsController_updateSubscription",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Organizations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/invoices": {
      "get": {
        "operationId": "InvoicesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/organizations/{orgId}/invoices": {
      "get": {
        "operationId": "InvoicesController_findByOrg",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InvoicesController_createManual",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Platform Invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/invoices/{id}/mark-paid": {
      "put": {
        "operationId": "InvoicesController_markPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkPaidDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/platform/invoices/{id}/void": {
      "put": {
        "operationId": "InvoicesController_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Platform Invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "IGreen HRMS API",
    "description": "Enterprise Human Resource Management System API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "pattern": "/^\\d{10}$/"
          }
        }
      },
...
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "maxEmployees": {
            "type": "number"
          },
          "priceMonthly": {
            "type": "number"
          },
          "priceQuarterly": {
            "type": "number"
          },
          "priceYearly": {
            "type": "number"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "maxEmployees",
          "priceMonthly",
          "priceQuarterly",
          "priceYearly"
        ]
      },
      "RegisterOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "pattern": "/^\\d{10}$/"
          },
          "planId": {
            "type": "string"
          },
          "billingCycle": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "QUARTERLY",
              "YEARLY"
            ]
          },
          "gracePeriodDays": {
            "type": "number"
          },
          "adminEmail": {
            "type": "string"
          },
          "adminName": {
            "type": "string"
          },
          "taxPercent": {
            "type": "number"
          },
          "dueAfterDays": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "email",
          "planId",
          "billingCycle",
          "adminEmail"
        ]
      },
      "UpdateSubscriptionDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string"
          },
          "billingCycle": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "QUARTERLY",
              "YEARLY"
            ]
          }
        }
      },
      "MarkPaidDto": {
        "type": "object",
        "properties": {
          "paymentDate": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "referenceNumber": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "recordedById": {
            "type": "string"
          }
        },
        "required": [
          "paymentDate",
          "paymentMethod",
          "amount",
          "recordedById"
        ]
      },
...
    }
  }
}